+++ Assets/Scripts/Projectile.cs
@@ -4,7 +4,7 @@ using UnityEngine.Assertions;
 public class Projectile : MonoBehaviour
 {
     [SerializeField] private float lifetime = 5.0f;
-    [SerializeField] private float damage = 5.0f;
+    [SerializeField] private int damage = 5;
     private bool _active = true;
 
     private void Awake()
